+++ ios/Podfile
@@ -28,6 +28,8 @@ target 'TravelDiary' do
   pod 'glog', :podspec => '../node_modules/react-native/third-party-podspecs/glog.podspec'
   pod 'Folly', :podspec => '../node_modules/react-native/third-party-podspecs/Folly.podspec'
 
+  pod 'RNGestureHandler', :path => '../node_modules/react-native-gesture-handler'
+
   target 'TravelDiaryTests' do
     inherit! :search_paths
     # Pods for testing
