+++ package.json
@@ -10,13 +10,15 @@
   "dependencies": {
     "firebase": "^6.3.0",
     "react": "16.8.6",
-    "react-native": "0.60.3",
+    "react-native": "^0.60.3",
+    "react-native-gesture-handler": "^1.3.0",
     "react-navigation": "^3.11.1"
   },
   "devDependencies": {
     "@babel/core": "7.5.4",
     "@babel/runtime": "7.5.4",
     "@react-native-community/eslint-config": "0.0.3",
+    "@types/react-navigation": "^3.0.7",
     "babel-jest": "24.8.0",
     "eslint": "6.0.1",
     "jest": "24.8.0",
